wip: implementing user sign up functionality  -making sure user leave no input on enterd
