+++ index.html
@@ -65,6 +65,7 @@
             <div>
                 <label for="firstName">First Name</label>
                 <input type="text" id="firstname" placeholder="INPUT YOUR FIRST NAME">
+                <p class="errorMes">error</p>
             </div>
             <div>
                 <label for="firstName">Last Name</label>
@@ -80,7 +81,7 @@
             </div>
             <div>
                 <label for="firstName">Password</label>
-                <input type="password" id="password" placeholder="CREATE PASSWORD">
+                <input type="password" id="password"  placeholder="CREATE PASSWORD">
             </div>
             <div>
                 <label for="firstName">Confirm Password</label>
@@ -546,6 +547,6 @@
     </div>
     </section>
    </main>
- <script src="index.js"></script>   
+ <script type="module" src="index.js"></script>   
 </body>
 </html>
